Added a loader in place of the refresh button as discussed in #141
This looks a lot better than the default infinite progressbar in ABS,
as it appears on the action item. This has a small downside when in
landscape though. As it can't show the text when it has an ActionView,
in some cases it'll have 'enough room' to show both the export and import
action items. Causing the account picker to be a tad smaller in some cases.
In my opinion the import and export shouldn't even be in the ActionBar as
they aren't exactly features a user uses a lot.