A basic implementation of JB notifications
All it really does now is show the normal text in a BigText, meaning that
if it's long and isn't fully readable you can expand it (on JB that is) to
read everything. It'll also show the account name it's for in the summary.
I also re-enabled the writing to the database so everyone can really test
this build without it not working properly. Note that this is just the
beginning, I'm planning on doing a lot more with this.