Small inconsistency: Removed generic Intent i and moved dependencies to ...
As commented by : @nelenkov on prior patch: https://github.com/AndlyticsProject/andlytics/pull/480
"Looks good, but can you please make this against the dev branch? We try not to touch master unless something in the API breaks. It will be merged into master in the next release."


As commented by : @nelenkov on prior patch: https://github.com/AndlyticsProject/andlytics/pull/479
"Thanks for the patch. However, I don't think this makes it more readable. Getting rid of the Intent i would be better IMHO. It's not like the switch just sets up the intent and it is started at the end of the method -- some cases use startActivity() and others startActivityForResult() so it is better each case to use an explicit, named intent."